+++ src/server/posts/getPosts.ts
@@ -1,6 +1,8 @@
+import { z } from 'zod';
 import { Post } from '@typings/posts/post';
+import { PostSchema } from '@schemas/posts/post.schema';
 
-export async function getPosts(limit?: number, category?: number, tag?: number, author?: number, search?: string): Promise<Post[] | undefined> {
+export async function getPosts(limit?: number, category?: number, tag?: number, author?: number, search?: string): Promise<Post[]> {
   try {
     const res = await fetch('/data/posts.json');
 
@@ -8,32 +10,38 @@ export async function getPosts(limit?: number, category?: number, tag?: number, 
       throw new Error(`Failed to fetch posts.json: ${res.status}`);
     }
 
-    const posts: Post[] = await res.json();
+    const unparsed = await res.json();
 
-    let filteredPosts = posts;
+    const parsed = z.array(PostSchema).safeParse(unparsed);
+
+    if (!parsed.success) {
+      throw new Error(`Invalid data: ${parsed.error}`);
+    }
+
+    let posts = parsed.data;
 
     if (limit && limit > 0) {
-      filteredPosts = filteredPosts.slice(0, limit);
+      posts = posts.slice(0, limit);
     }
 
     if (category) {
-      filteredPosts = filteredPosts.filter((post) => post.categories?.includes(category));
+      posts = posts.filter((post) => post.categories?.includes(category));
     }
 
     if (tag) {
-      filteredPosts = filteredPosts.filter((post) => post.tags?.includes(tag));
+      posts = posts.filter((post) => post.tags?.includes(tag));
     }
 
     if (author) {
-      filteredPosts = filteredPosts.filter((post) => post.authorId === author);
+      posts = posts.filter((post) => post.authorId === author);
     }
 
     if (search) {
       const lowerSearch = search.toLowerCase();
-      filteredPosts = filteredPosts.filter((post) => post.title.toLowerCase().includes(lowerSearch));
+      posts = posts.filter((post) => post.title.toLowerCase().includes(lowerSearch));
     }
 
-    return filteredPosts;
+    return posts;
   } catch (error) {
     console.error('getPosts', error);
     throw error;