+++ src/server/products/getTags.ts
@@ -1,6 +1,8 @@
+import { z } from 'zod';
 import { Tag } from '@typings/products/tag';
+import { TagSchema } from '@schemas/posts/tag.schema';
 
-export async function getTagsArray(tagIds: number[]): Promise<Tag[] | undefined> {
+export async function getTagsArray(tagIds: number[]): Promise<Tag[]> {
   try {
     const res = await fetch('/data/product-tags.json');
 
@@ -8,7 +10,15 @@ export async function getTagsArray(tagIds: number[]): Promise<Tag[] | undefined>
       throw new Error(`Failed to fetch product-tags.json: ${res.status}`);
     }
 
-    const tags: Tag[] = await res.json();
+    const unparsed = await res.json();
+
+    const parsed = z.array(TagSchema).safeParse(unparsed);
+
+    if (!parsed.success) {
+      throw new Error(`Invalid data: ${parsed.error}`);
+    }
+
+    const tags = parsed.data;
     const idSet = new Set(tagIds);
     const tagArray = tags.filter((tag) => idSet.has(tag.id));