+++ src/server/shop/getCustomer.ts
@@ -1,4 +1,6 @@
+import { z } from 'zod';
 import { Customer } from '@typings/shop/customer';
+import { CustomerSchema } from '@schemas/shop/customer.schema';
 
 export async function getCustomer<K extends 'id' | 'email'>(field: K, value: Customer[K]): Promise<Customer | undefined> {
   try {
@@ -8,7 +10,15 @@ export async function getCustomer<K extends 'id' | 'email'>(field: K, value: Cus
       throw new Error(`Failed to fetch shop-customers.json: ${res.status}`);
     }
 
-    const customers: Customer[] = await res.json();
+    const unparsed = await res.json();
+
+    const parsed = z.array(CustomerSchema).safeParse(unparsed);
+
+    if (!parsed.success) {
+      throw new Error(`Invalid data: ${parsed.error}`);
+    }
+
+    const customers = parsed.data;
     const customer = customers.find((customer) => customer[field] === value);
 
     return customer;