+++ src/server/shop/getOrder.ts
@@ -1,4 +1,6 @@
+import { z } from 'zod';
 import { Order } from '@typings/shop/order';
+import { OrderSchema } from '@schemas/shop/order.schema';
 
 export async function getOrder<K extends 'id'>(field: K, value: Order[K]): Promise<Order | undefined> {
   try {
@@ -8,7 +10,15 @@ export async function getOrder<K extends 'id'>(field: K, value: Order[K]): Promi
       throw new Error(`Failed to fetch shop-orders.json: ${res.status}`);
     }
 
-    const orders: Order[] = await res.json();
+    const unparsed = await res.json();
+
+    const parsed = z.array(OrderSchema).safeParse(unparsed);
+
+    if (!parsed.success) {
+      throw new Error(`Invalid data: ${parsed.error}`);
+    }
+
+    const orders = parsed.data;
     const order = orders.find((order) => order[field] === value);
 
     return order;