+++ src/server/shop/getOrders.ts
@@ -1,6 +1,8 @@
+import { z } from 'zod';
 import { Order } from '@typings/shop/order';
+import { OrderSchema } from '@schemas/shop/order.schema';
 
-export async function getOrders<K extends 'customerId'>(field: K, value: Order[K]): Promise<Order[] | undefined> {
+export async function getOrders<K extends 'customerId'>(field: K, value: Order[K]): Promise<Order[]> {
   try {
     const res = await fetch('/data/shop-orders.json');
 
@@ -8,7 +10,15 @@ export async function getOrders<K extends 'customerId'>(field: K, value: Order[K
       throw new Error(`Failed to fetch shop-orders.json: ${res.status}`);
     }
 
-    const ordersList: Order[] = await res.json();
+    const unparsed = await res.json();
+
+    const parsed = z.array(OrderSchema).safeParse(unparsed);
+
+    if (!parsed.success) {
+      throw new Error(`Invalid data: ${parsed.error}`);
+    }
+
+    const ordersList = parsed.data;
     const orders = ordersList.filter((order) => order[field] === value);
 
     return orders;