+++ src/types/products/product.ts
@@ -10,12 +10,12 @@ export interface Product {
   image?: string;
   gallery?: string[];
   shortDescription?: string;
-  description?: ContentBlock;
+  description?: ContentBlock[];
   price: number;
   inStock: boolean;
   SKU: string;
   variationAttributes?: { type: string; options: number[] }[];
-  variations?: { id: number; sizeId: number; price: number; stock: number; SKU: string }[];
+  variations?: { id: number; colourId?: number; sizeId?: number; price: number; stock: number; SKU: string }[];
   reviewCount?: number;
   averageReview?: number;
   relatedProducts?: number[];